Add unittest for codefile_class
